# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "Simple Crud Api product",
    "title": "ITOMMEY API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/product": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "List all available product",
        "description": "Description for list product",
        "operationId": "getProducts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Add new Product",
        "description": "Create new",
        "operationId": "addProduct",
        "requestBody": {
          "$ref": "#/components/requestBodies/Product"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Product"
          },
          "400": {
            "description": "Failed to create new Product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Failed to create new Porduct"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "lifeboy"
          },
          "qty": {
            "type": "number",
            "example": 10
          },
          "picture": {
            "type": "string",
            "example": "base64file"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "expiredAt": {
            "type": "string",
            "example": "2022-11-17 04:33:12"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-9-12 04:33:12"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-9-12 04:33:12"
          }
        }
      }
    },
    "requestBodies": {
      "Product": {
        "description": "Product object",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "lifeboy"
                },
                "qty": {
                  "type": "number",
                  "example": 10
                },
                "picture": {
                  "type": "string",
                  "example": "base64file"
                },
                "expiredAt": {
                  "type": "string",
                  "example": "2022-11-17 04:33:12"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "Product": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "lifeboy"
                },
                "qty": {
                  "type": "number",
                  "example": 10
                },
                "picture": {
                  "type": "string",
                  "example": "base64file"
                },
                "isActive": {
                  "type": "boolean",
                  "example": true
                },
                "expiredAt": {
                  "type": "string",
                  "example": "2022-11-17 04:33:12"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2022-9-12 04:33:12"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2022-9-12 04:33:12"
                }
              }
            }
          }
        }
      }
    },
...
  }
}